+++ build.gradle
@@ -57,6 +57,8 @@ dependencies {
 
 	// smtp
 	implementation 'org.springframework.boot:spring-boot-starter-mail'
+	// thymeleaf
+	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
 
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
