+++ src/main/java/com/idear/backend/email/service/EmailService.java
@@ -1,0 +1,52 @@
+package com.idear.backend.email.service;
+
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
+import lombok.RequiredArgsConstructor;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.stereotype.Service;
+import org.thymeleaf.TemplateEngine;
+import org.thymeleaf.context.Context;
+
+import java.util.Map;
+
+import com.idear.backend.global.exception.CustomException;
+import com.idear.backend.global.exception.ErrorCode;
+
+@Service
+@RequiredArgsConstructor
+public class EmailService {
+
+    private final JavaMailSender mailSender;
+    private final TemplateEngine templateEngine;
+
+    public void sendEmail(String email, String title, String text) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(email);
+        message.setSubject(title);
+        message.setText(text);
+        mailSender.send(message);
+    }
+
+    public void sendEmailWithTemplate(String email, String title, String templateName, Map<String, Object> variables) {
+        Context context = new Context();
+        context.setVariables(variables);
+
+        String htmlContent = templateEngine.process("email/" + templateName, context);
+
+        try {
+            MimeMessage mimeMessage = mailSender.createMimeMessage();
+            MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, true, "UTF-8");
+
+            helper.setTo(email);
+            helper.setSubject(title);
+            helper.setText(htmlContent, true);
+
+            mailSender.send(mimeMessage);
+        } catch (MessagingException e) {
+            CustomException.of(ErrorCode.EMAIL_SEND_FAILED);
+        }
+    }
+}
