--- src/main/java/com/idear/backend/global/infrasturcture/EmailService.java
@@ -1,22 +1,0 @@
-package com.idear.backend.global.infrasturcture;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-
-
-@Service
-@RequiredArgsConstructor
-public class EmailService {
-
-    private final JavaMailSender mailSender;
-
-    public void sendEmail(String email, String title, String text) {
-        SimpleMailMessage message = new SimpleMailMessage();
-        message.setTo(email);
-        message.setSubject(title);
-        message.setText(text);
-        mailSender.send(message);
-    }
-}
